Work to get functional dequant-matmul on Coop matrix.
1. CollapseNestedTileForOps - Remove sunkenOps check to handle cases where no
  ops in first/parent loop
2. Modify sharedMemory to take into account current vectorSize before
   re-adjusting vectorSize to be bigger than necessary for sub-bytes
stuff. Did this to have more similar tile/unroll size to regular matmul.
3. Multibuffering added some gpu.barrier to synchronize before and after dequant ops. (more similar to matmul)
4. On GPUPipelining, this is main fix for correctness, add support in
   cases where transferReadOp do not directly feed into transferWrite.
for example in dequant case we'd need to feed to the dequant operation +
since it is elementwise operation(with max vector size of 4)  we need to introduce
extract_strided_slice and insert_strided_slice to handle our
pre-selected larger vector size of 8 or 16.
5.Added comment/todo to handle defaultNativeShape for transferWrite
  better, current way would not fold transfer_read-transfer_write properly.