[GPU][VectDist] Refactor multiReducOp lowering to reduce acc at the end.
The main motivation behind this commit is to fix the numerics issue we
find in Attention GPU C++ pipeline. Where our output seems to be some
scale off from reference (i.e out = k * ref). Through experiments we
determine that the cause of the issue is the multiReducOp distribution,
which is required for scaling/merge computation.

Through analyzing of IR and experimentation, reducing the srcVector and
non constant accumulator at the same time with another multiDimReduction
seems to output numerically wrong values.

This commit fixes numerical issue by refactoring the distribution of
multireductionOp to 3 steps. First every thread locally reduce the srcVector
it holds with combiningIdentity as localInit. Second, it does a subgroup/warp
reduce amongs other threads. Finally, each thread does a local reduction of the
intermediate reduced data it has with the accumulator it holds.

Signed-off-by: Stanley Winata <stanley.winata@amd.com>