Optimize `fp8` `linalg_ext.attention` using Q@K offset
Rather than scaling `Q@K` post 	`exp2` to adjust for `fp8` range we
can apply a linear offset to the initial matmul. This suses `max` to
have the same offset (cancelled out during iteration) and sum to be
scaled by the same amount. The outcome is that the same numerical
outcome with fewer operations.